home page controller added
